backend/requirements.txt
django
djangorestframework
django-cors-headers
httpx
python-dotenv
Pillow
gunicorn
